Fixed issues: -added minicart -fixed bigdecimal in recalculate method

+++ src/main/java/com/es/phoneshop/web/cart/DeleteCartItemServlet.java
@@ -9,6 +9,7 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import org.apache.maven.shared.utils.StringUtils;
 
 import java.io.IOException;
 
@@ -24,16 +25,19 @@ public class DeleteCartItemServlet extends HttpServlet {
 
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        long productId = Long.valueOf(request.getPathInfo().substring(1));
+        if (!StringUtils.isEmpty(request.getPathInfo())) {
+            long productId = Long.valueOf(request.getPathInfo().substring(1));
 
-        Cart cart = cartService.getCart(request);
-        try {
-            cartService.deleteProductInCart(cart, productId);
-        } catch (ProductNotFoundException e) {
-            response.sendRedirect(request.getContextPath() + "/cart?errors=" + e.getMessage());
+            Cart cart = cartService.getCart(request);
+            try {
+                cartService.deleteProductInCart(cart, productId);
+            } catch (ProductNotFoundException e) {
+                response.sendRedirect(request.getContextPath() + "/cart?errors=" + e.getMessage());
+            }
+            response.sendRedirect(request.getContextPath() + "/cart?message=Cart item removed successfully");
+        } else {
+            response.sendError(HttpServletResponse.SC_NOT_FOUND, "Page not found");
         }
-
-        response.sendRedirect(request.getContextPath() + "/cart?message=Cart item removed successfully");
     }
 
 }
